restored info.go, cd.go, added internals.go
